bring the seed file formatting back into line with Rails
Rails treats the seed file as a simple script. There is no module/class
structure to organize and no method to run. Simply requiring it should
be enough to seed the database.

Also, seed files should use `#create!` so that they raise an error if
there was an issue saving.